Added in scaling ability to Frame2D and also fixed a bug in polygons going off screen causing an error.

+++ src/engine2d/RenderingEngine2D.h
@@ -49,17 +49,17 @@ namespace softengine
 		Color refreshColor;
 		bool isInitialised;
 
-		Color GetPixelValue(size_t pixel) const;
+		Color GetPixelValue(int pixel) const;
 		Color GetPixelValue(
-			size_t pixelX,
-			size_t pixelY) const;
+			int pixelX,
+			int pixelY) const;
 		void SetPixelValue(
-			size_t pixel,
+			int pixel,
 			Color& color
 		);
 		void SetPixelValue(
-			size_t pixelX,
-			size_t pixelY,
+			int pixelX,
+			int pixelY,
 			Color& color
 		);
