+++ src/engine2d/Utils2D.cpp
@@ -2,22 +2,22 @@
 
 using namespace softengine;
 
-std::vector<size_t> Utils2D::InterpolateXPixelValues(
-	size_t x0,
-	size_t y0,
-	size_t x1,
-	size_t y1)
+std::vector<int> Utils2D::InterpolateXPixelValues(
+	int x0,
+	int y0,
+	int x1,
+	int y1)
 {
 	return InterpolateYPixelValues(y0, x0, y1, x1);
 }
 
-std::vector<size_t> Utils2D::InterpolateYPixelValues(
-	size_t x0,
-	size_t y0,
-	size_t x1,
-	size_t y1)
+std::vector<int> Utils2D::InterpolateYPixelValues(
+	int x0,
+	int y0,
+	int x1,
+	int y1)
 {
-	std::vector<size_t> pixels;
+	std::vector<int> pixels;
 
 	if (x0 == x1)
 	{
@@ -29,9 +29,9 @@ std::vector<size_t> Utils2D::InterpolateYPixelValues(
 	double grad = ((double)y1 - (double)y0) / ((double)x1 - (double)x0);
 	double d = (double)y0;
 
-	for (size_t i = x0; i <= x1; i++)
+	for (int i = x0; i <= x1; i++)
 	{
-		pixels.push_back((size_t)d);
+		pixels.push_back(d);
 		d += grad;
 	}
