+++ src/engine2d/Utils2D.h
@@ -9,18 +9,18 @@ namespace softengine
 	public:
 		Utils2D() = delete;
 
-		static std::vector<size_t> InterpolateXPixelValues(
-			size_t x0,
-			size_t y0,
-			size_t x1,
-			size_t y1
+		static std::vector<int> InterpolateXPixelValues(
+			int x0,
+			int y0,
+			int x1,
+			int y1
 		);
 
-		static std::vector<size_t> InterpolateYPixelValues(
-			size_t x0,
-			size_t y0,
-			size_t x1,
-			size_t y1
+		static std::vector<int> InterpolateYPixelValues(
+			int x0,
+			int y0,
+			int x1,
+			int y1
 		);
 	};
 }
